modularize code
